TMP107 estop condition cheat method complete

+++ sensors/sensor_api.h
@@ -47,7 +47,7 @@ float getPower();
 uint8_t getAcceleration();
 
 // Update threshold values that trigger an eStop
-void setThresholdTemp(uint8_t thresholdTemp); // Use Celsius
+void setThresholdTemp(uint16_t thresholdTemp); // Use Celsius
 void setThresholdCurrent(uint16_t thresholdCurrent); // Use mA
 void setThresholdAccel(uint16_t thresholdAccel); // Use m/s^2
